Add working web router
