+++ transition/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import configureStore from './store/store';
-import RootContainer from './root';
+import RootContainer from './components/root';
 import { getAuthToken } from './util/mobile_auth_util';
 import { fetchCurrentSession } from './actions/session_actions';
