+++ transition/src/components/home/home.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, Button, Text } from 'react-native';
-import styles from '../../styles';
+import styles from '../../styles/styles';
 
 const Home = ({ getPublicDoc, userId, createDoc, logout }) => (
   <View style={styles.container}>
