+++ transition/src/components/splash/splash.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View } from 'react-native';
 import AuthContainer from './auth_container';
-import styles from '../../styles';
+import styles from '../../styles/styles';
 
 const Splash = () => (
   <View style={styles.container}>
