+++ transition/src/store/store.js
@@ -1,7 +1,6 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
-import differ from './differ';
 import RootReducer from '../reducers/root_reducer';
 
 const configureStore = (preloadedState = {}) => (
@@ -10,7 +9,6 @@ const configureStore = (preloadedState = {}) => (
     preloadedState,
     applyMiddleware(
       thunk,
-      differ,
       logger,
     )
   )
