+++ transition/src/util/platform_util.js
@@ -1,0 +1,9 @@
+const navByPlatform = (ownProps, mobileLocation, webLocation) => {
+  if (ownProps.navigation) {
+    ownProps.navigation.navigate(mobileLocation);
+  } else {
+    ownProps.history.push(webLocation);
+  }
+};
+
+export default navByPlatform;
